Clear event listeners on thresher after each URL (fixes #33)
When the URLlist is iterated over using recursive setTimeOuts,
the Thresher object is staying in scope for all the nested calls.

This is leading to event listeners accumulating on the Thresher
object, and multiple identical handlers being called for the same
event. As a shim, I have simply cleared all the event listeners
after each URL finishes processing.